refactor: 패키지 수정 및 반영
